--- a/pandas/core/dtypes/dtypes.py
+++ b/pandas/core/dtypes/dtypes.py
@@ -6,7 +6,6 @@ import pytz
 
 from pandas._libs.interval import Interval
 from pandas._libs.tslibs import NaT, Period, Timestamp, timezones
-from pandas.util._decorators import cache_readonly
 
 from pandas.core.dtypes.generic import ABCCategoricalIndex, ABCIndexClass
 
@@ -532,12 +531,12 @@ class DatetimeTZDtype(PandasExtensionDtype):
         self._unit = unit
         self._tz = tz
 
-    @cache_readonly
+    @property
     def unit(self):
         """The precision of the datetime data."""
         return self._unit
 
-    @cache_readonly
+    @property
     def tz(self):
         """The timezone."""
         return self._tz