--- a/pandas/core/dtypes/dtypes.py
+++ b/pandas/core/dtypes/dtypes.py
@@ -1,5 +1,6 @@
 """ define extension dtypes """
 import re
+import warnings
 
 import numpy as np
 import pytz
@@ -483,7 +484,6 @@ class DatetimeTZDtype(PandasExtensionDtype):
     str = '|M8[ns]'
     num = 101
     base = np.dtype('M8[ns]')
-    na_value = NaT
     _metadata = ('unit', 'tz')
     _match = re.compile(r"(datetime64|M8)\[(?P<unit>.+), (?P<tz>.+)\]")
     _cache = {}
@@ -517,7 +517,20 @@ class DatetimeTZDtype(PandasExtensionDtype):
             unit, tz = unit.unit, unit.tz
 
         if unit != 'ns':
-            raise ValueError("DatetimeTZDtype only supports ns units")
+            if isinstance(unit, compat.string_types) and tz is None:
+                # maybe a string like datetime64[ns, tz], which we support for
+                # now.
+                result = type(self).construct_from_string(unit)
+                unit = result.unit
+                tz = result.tz
+                msg = (
+                    "Passing a dtype alias like 'datetime64[ns, {tz}]' "
+                    "to DatetimeTZDtype is deprecated. Use "
+                    "'DatetimeTZDtype.construct_from_string()' instead."
+                )
+                warnings.warn(msg.format(tz=tz), FutureWarning, stacklevel=2)
+            else:
+                raise ValueError("DatetimeTZDtype only supports ns units")
 
         if tz:
             tz = timezones.maybe_get_tz(tz)
@@ -567,7 +580,7 @@ class DatetimeTZDtype(PandasExtensionDtype):
         >>> DatetimeTZDtype.construct_from_string('datetime64[ns, UTC]')
         datetime64[ns, UTC]
         """
-        msg = "Could not construct DatetimeTZDtype from {}"
+        msg = "Could not construct DatetimeTZDtype from '{}'"
         try:
             match = cls._match.match(string)
             if match: