--- a/pandas/tests/dtypes/test_common.py
+++ b/pandas/tests/dtypes/test_common.py
@@ -42,9 +42,12 @@ class TestPandasDtype(object):
         'datetime64[ns, US/Eastern]',
         'datetime64[ns, Asia/Tokyo]',
         'datetime64[ns, UTC]'])
+    @pytest.mark.xfail(reason="dtype-caching", strict=True)
     def test_datetimetz_dtype(self, dtype):
-        assert com.pandas_dtype(dtype) is DatetimeTZDtype(dtype)
-        assert com.pandas_dtype(dtype) == DatetimeTZDtype(dtype)
+        assert (com.pandas_dtype(dtype) is
+                DatetimeTZDtype.construct_from_string(dtype))
+        assert (com.pandas_dtype(dtype) ==
+                DatetimeTZDtype.construct_from_string(dtype))
         assert com.pandas_dtype(dtype) == dtype
 
     def test_categorical_dtype(self):