@pep8speaks 2018-11-29
Hello @TomAugspurger! Thanks for submitting the PR.

 - There are no PEP8 issues in the file [`pandas/core/arrays/datetimelike.py`](https://github.com/TomAugspurger/pandas/blob/1ca7fa4fe611ccee04c657eae836049caef18f6c/pandas/core/arrays/datetimelike.py) !

 - There are no PEP8 issues in the file [`pandas/core/dtypes/common.py`](https://github.com/TomAugspurger/pandas/blob/1ca7fa4fe611ccee04c657eae836049caef18f6c/pandas/core/dtypes/common.py) !

 - There are no PEP8 issues in the file [`pandas/core/dtypes/dtypes.py`](https://github.com/TomAugspurger/pandas/blob/1ca7fa4fe611ccee04c657eae836049caef18f6c/pandas/core/dtypes/dtypes.py) !

 - There are no PEP8 issues in the file [`pandas/tests/dtypes/test_common.py`](https://github.com/TomAugspurger/pandas/blob/1ca7fa4fe611ccee04c657eae836049caef18f6c/pandas/tests/dtypes/test_common.py) !

 - There are no PEP8 issues in the file [`pandas/tests/dtypes/test_dtypes.py`](https://github.com/TomAugspurger/pandas/blob/1ca7fa4fe611ccee04c657eae836049caef18f6c/pandas/tests/dtypes/test_dtypes.py) !

 - There are no PEP8 issues in the file [`pandas/tests/dtypes/test_missing.py`](https://github.com/TomAugspurger/pandas/blob/1ca7fa4fe611ccee04c657eae836049caef18f6c/pandas/tests/dtypes/test_missing.py) !