@TomAugspurger 2018-11-29
On microbenchmarks, things are fine

master:

```python
In [2]: %time a = pd.core.dtypes.dtypes.DatetimeTZDtype(unit='ns', tz="UTC")
CPU times: user 39 µs, sys: 31 µs, total: 70 µs
Wall time: 75.6 µs

In [3]: %time a = pd.core.dtypes.dtypes.DatetimeTZDtype(unit='ns', tz="UTC")
CPU times: user 15 µs, sys: 0 ns, total: 15 µs
Wall time: 18.8 µs
```

PR:

```python
In [2]: %time a = pd.core.dtypes.dtypes.DatetimeTZDtype(unit='ns', tz="UTC")
CPU times: user 39 µs, sys: 28 µs, total: 67 µs
Wall time: 71 µs

In [3]: %time a = pd.core.dtypes.dtypes.DatetimeTZDtype(unit='ns', tz="UTC")
CPU times: user 15 µs, sys: 0 ns, total: 15 µs
Wall time: 19.3 µs
```



ASV for the timeseries, timestamps, and offsets

```
       before           after         ratio
     [6b3490f4]       [982c169a]
-         143±9ms          127±4ms     0.89  timeseries.DatetimeIndex.time_add_timedelta('tz_aware')
```