@TomAugspurger 2018-11-29
Are you worried about memory usage? The time to create a new one from
scratch is identical or faster than looking it up from the cache.

On Thu, Nov 29, 2018 at 7:06 AM Jeff Reback <notifications@github.com>
wrote:

> *@jreback* commented on this pull request.
> ------------------------------
>
> In pandas/core/dtypes/dtypes.py
> <https://github.com/pandas-dev/pandas/pull/23990#discussion_r237478340>:
>
> >      _metadata = ('unit', 'tz')
>      _match = re.compile(r"(datetime64|M8)\[(?P<unit>.+), (?P<tz>.+)\]")
>      _cache = {}
> +    # TODO: restore caching? who cares though? It seems needlessly complex.
> +    # np.dtype('datetime64[ns]') isn't a singleton
>
> try running a good part of the test suite. Its the repeated consruction
> that's a problem, not the single contruction which is fine. W/o caching you
> end up creating a huge number of these
>
> —
> You are receiving this because you modified the open/close state.
> Reply to this email directly, view it on GitHub
> <https://github.com/pandas-dev/pandas/pull/23990#discussion_r237478340>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABQHIuydvXMpt0620svkG4I2c9VNuJfMks5uz9vOgaJpZM4Y42qg>
> .
>