@jreback 2018-11-29
maybe things changed since I created this originally, but w/o caching perf was not great.