@TomAugspurger 2018-11-29
I've removed `_coerce_to_dtype`.