@jreback 2018-11-29
dont you need to remove the ``_coerce_to_dtype`` tests?