@TomAugspurger 2018-11-29
https://github.com/pandas-dev/pandas/pull/23990/commits/7ab2a74c304c6791f8f28edc01b7f43b2cbd2fab removed all the `_coerce_to_dtype` tests.