@TomAugspurger 2018-11-30
I may be doing something wrong, but it seems like memory usage isn't increasing either.

```python
from memory_profiler import profile
import pandas as pd


@profile(precision=4)
def check():
    dts = [pd.core.dtypes.dtypes.DatetimeTZDtype('ns', 'UTC')] * 100_000_000
    return dts
```

master: ![master](https://user-images.githubusercontent.com/1312546/49298301-a078d780-f482-11e8-9ac9-963dad296d60.png)


pr: ![pr](https://user-images.githubusercontent.com/1312546/49298302-a078d780-f482-11e8-9a63-4db09cbc7ead.png)

But the chart for master doesn't look right. Shouldn't that be linear in the number of DatetimeTZDtypes we create? Perhaps it is correctly, and we're just not reporting our memory usage correctly?