@TomAugspurger 2018-11-30
Although memory_profiler uses psutil, which I don't *think* relies on `__sizeof__`, though I may be wrong.