@TomAugspurger 2018-12-03
I'll revert the magic construction changes with a futurewarning. But DatetimeTZDtype isn't part of the public API. Dask, at least, doesn't use it directly.