@TomAugspurger 2018-12-03
Added to the deprecation log.