@jreback 2018-12-03
> Do we typically add release notes for something like this (deprecations that aren't part of the public API, but are possibly in use downstream)?

this IS certainly part of the public API as ``DatetimeTZDtype`` is public.