@jreback 2018-12-03
I *know* ``pyarrow`` uses this at the very least. (I don't know if the 2 form of the construction is used though). cc @wesm