@TomAugspurger 2018-12-03
Added a release note. Going to followup on api.rst later, since we don't have an `array` section for datetime array there yet.