@datapythonista 2018-12-03
Haven't seen those errors before, I guess something is failing in their side. @DavidStaheli do you mind taking a look at https://dev.azure.com/pandas-dev/pandas/_build/results?buildId=4597