@jreback 2018-12-03
@TomAugspurger needs a rebase as well