@vtbassmatt 2018-12-03
I suspect it's a little bit of abuse fighting that was turned up a little too high. I think we've addressed it now.